cube: add glowing mode & comments
The mode "glowing" has been added onto the program, it will randomely
pick an intensity and make the whole strip LED brighter until that
point; and then reduce it back to zero. And so on and so on.

Code comments have also been added to easer reading/understanding based
on the different mode existing.